Introduce the record header class
Currently, the header contains only the record number. In asynchronous applications with multiple queues, one would like to know the data source from which the record has been read as well as the timestamp.

+++ easybatch-core/src/main/java/org/easybatch/core/api/RecordReader.java
@@ -56,7 +56,7 @@ public interface RecordReader {
      * @return the total record number in the data source or null if the total records number cannot be
      * calculated in advance
      */
-    Integer getTotalRecords();
+    Long getTotalRecords();
 
     /**
      * This method returns a human readable data source name to be shown in the batch report.
