+++ easybatch-core/src/main/java/org/easybatch/core/impl/NoOpRecordReader.java
@@ -58,7 +58,7 @@ class NoOpRecordReader implements RecordReader {
     /**
      * {@inheritDoc}
      */
-    public Integer getTotalRecords() {
+    public Long getTotalRecords() {
         return null;
     }
